+++ flecsi/exec/leg/tracer.hh
@@ -2,9 +2,9 @@
 #define FLECSI_LEG_EXEC_TRACER_HH
 
 #include "flecsi/flog.hh"
+#include "flecsi/run/backend.hh"
 #include "flecsi/util/common.hh"
-
-#include <legion.h>
+#include "flecsi/util/types.hh" // Color
 
 namespace flecsi::exec {
 
@@ -15,39 +15,72 @@ struct trace {
 
   inline guard make_guard();
 
-  trace() : id_(Legion::Runtime::get_runtime()->generate_dynamic_trace_id()) {}
-  [[deprecated("use default constructor")]] explicit trace(id_t id) : id_(id) {}
+  trace()
+    : dat(std::make_unique<data>(
+        Legion::Runtime::get_runtime()->generate_dynamic_trace_id())) {}
+  [[deprecated("use default constructor")]] explicit trace(id_t id)
+    : dat(std::make_unique<data>(id)) {}
 
   void skip() {
     skip_ = true;
   }
 
 private:
+  struct data {
+    explicit data(id_t id) : id(id) {}
+    data(data &&) = delete; // for (cautious) address stability
+
+    operator id_t() const {
+      return id;
+    }
+
+    void rewind() {
+      where = 0;
+    }
+    const run::task_count::ptr & next() {
+      return where++ < tasks.size()
+               ? tasks[where - 1]
+               : tasks.emplace_back(std::make_shared<run::task_count>());
+    }
+
+  private:
+    id_t id;
+    std::vector<run::task_count::ptr> tasks; // for each launch in order
+    run::task_idx where = 0;
+  };
+
   void start() {
     if(!skip_) {
       if(tracing)
         flog_fatal("Trace already running: traces cannot be overlapping");
+      tracing = dat.get();
       // Call Legion tracing tool
       Legion::Runtime::get_runtime()->begin_trace(
         Legion::Runtime::get_context(),
-        id_.value(),
+        *tracing,
         false, // logical_only = false
         false, // static_trace  = false
         NULL // std::set<RegionTreeID> *managed = NULL
       );
-      tracing = true;
     }
   }
 
   void stop() {
     if(!skip_) {
+      flog_assert(tracing == dat.get(), "wrong trace");
       Legion::Runtime::get_runtime()->end_trace(
-        Legion::Runtime::get_context(), id_.value());
-      tracing = false;
+        Legion::Runtime::get_context(), *tracing);
+      tracing->rewind();
+      tracing = nullptr;
     }
     else {
       skip_ = false;
     }
+  }
+
+public:
+  static data * current() {
+    return tracing;
   }
 
   friend bool is_tracing() {
@@ -55,9 +88,9 @@ private:
   }
 
 private:
-  util::move_optional<id_t> id_;
+  std::unique_ptr<data> dat;
   bool skip_ = false;
-  static inline bool tracing = false;
+  static inline data * tracing = nullptr;
 
 }; // struct trace