+++ flecsi/execution.hh
@@ -99,15 +99,7 @@ reduce(Args &&... args) {
   framework. Test tasks must return an integer that is non-zero on failure,
   and zero otherwise.
 
-  @tparam TASK       The user task. Its parameters may be of any
-                     default-constructible, trivially-move-assignable,
-                     non-pointer type, any type that supports the Legion
-                     return-value serialization interface, or any of several
-                     standard containers of such types. If \a ATTRIBUTES
-                     specifies an MPI task, parameters need merely be movable.
-  @tparam ATTRIBUTES The task attributes mask.
-  @tparam ARGS       The user-specified task arguments, implicitly converted to
-                     the parameter types for \a TASK.
+  \see \c execute about parameter and argument types.
 
   @return zero on success, non-zero on failure.
  */